Apply Jazzy-specific improvements and best practices
- Update C++ standard from C++14 to C++17 in all CMakeLists.txt (13 packages)
- Update CMake minimum version from 3.5 to 3.8 (14 packages)
- Modernize parameter API: get_parameter(name, var) -> get_parameter(name).as_<type>()
  - Updated parameters package to use as_string() and as_int()
  - Updated dynamic_tf2_publisher to use modern API with as_double()
  - Changed parameter types from float to double for better precision
- Update documentation (JAZZY_MIGRATION_NOTES.md and README.md)
- All packages now follow ROS 2 Jazzy best practices and recommendations

+++ CPP/dynamic_tf2_publisher/include/dynamic_tf2_publisher/tf2_publisher.h
@@ -27,7 +27,7 @@ public:
 
 private:
   std::string parentFrame_, childFrame_;
-  float xPos_, yPos_, zPos_, roll_, pitch_, yaw_;
+  double xPos_, yPos_, zPos_, roll_, pitch_, yaw_;
   int loopRate_;
   tf2::Quaternion quat_;
   std::unique_ptr<tf2_ros::TransformBroadcaster> tfBroadcaster_;
